+++ CHANGELOG.md
@@ -32,6 +32,7 @@ We haven't had the best track record of code/feature changes before this date, b
      Unreleased changes that have landed in master. Click to see more.
     </summary>
 
+- [feat] Remove Banzai & FbtLogger, provide hooks in their place.
 - [refactor] Abstract away string serialization error handling.
 - [refactor] Move FbtContentItem and $NestedFbtContentItems to libdefs
 - [refactor] Flow type strict substituteTokens.js